+++ aqua/core/analysis/analysis.py
@@ -11,7 +11,7 @@ import tempfile
 from importlib import resources as pypath
 
 from aqua.core.configurer import ConfigPath
-from aqua.core.dask.daskcluster import DaskCluster
+from aqua.core.dask import DaskCluster
 from aqua.core.logger import log_configure
 from aqua.core.util import create_folder, dump_yaml, format_realization, get_arg, load_yaml, to_list